# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar papéis",
                "description": "Retorna uma lista de papéis (roles) com suporte a paginação e filtros opcionais.",
                "operationId": "e4a0ea20b57f19c5b6d04c7c40f09173",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Quantidade de itens por página.",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar resultados pelo nome do papel.",
                        "schema": {
                            "type": "string",
                            "example": "Admin"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtrar por status ativo (true) ou inativo (false).",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de papéis retornada com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Criar um novo papel",
                "description": "Armazena um novo papel no sistema. É necessário fornecer um nome único e opcionalmente associar permissões.",
                "operationId": "6224ddd585cdd46a7054e7de33f3854c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do papel",
                                        "type": "string",
                                        "example": "Manager"
                                    },
                                    "active": {
                                        "description": "Status do papel",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "permissions": {
                                        "description": "Lista de UUIDs das permissões associadas.",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Papel criado com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Exibir detalhes de um papel",
                "description": "Retorna informações detalhadas de um papel específico identificado pelo seu UUID.",
                "operationId": "3122dc977d1062deb54c31ed424af409",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID do papel a ser recuperado.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do papel retornados com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Papel não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Atualizar um papel existente",
                "description": "Atualiza informações de um papel específico, como nome, status e permissões associadas.",
                "operationId": "349dad47b79d5a48712423dfc046506b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID do papel a ser atualizado.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do papel",
                                        "type": "string",
                                        "example": "Editor"
                                    },
                                    "active": {
                                        "description": "Status do papel",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "permissions": {
                                        "description": "Lista de UUIDs das permissões associadas.",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Papel atualizado com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Papel não encontrado."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Excluir um papel",
                "description": "Remove um papel do sistema identificado pelo seu UUID.",
                "operationId": "9056c9b2c62085b4af3271ecb91c301d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID do papel a ser excluído.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Papel excluído com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Papel não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}